Update VCR v0.1.0: Rename vtr→vcr, add benchmarks, update docs
- Renamed all vtr references to vcr across docs and code
- Fixed benchmark type errors (u32→usize for ByteRange)
- Added benchmark_runner.py for performance testing
- Added BENCHMARK_RESULTS.md with real metrics
- Updated README with improved architecture and usage
- Added Figure 3 (CPG structure) and Figure 5 (benchmarks) PNGs
- Updated snapshot path from vtr-snapshot to vcr-snapshot
- Updated all documentation (CLI_SCHEMAS, CI_INTEGRATION, etc.)
- Added ARCHITECTURE_IMPROVED.md with 4 diagram versions

Real benchmark results:
- Ingestion: 13.69ms (cold), 7.19ms (warm)
- Snapshot save/verify: ~7ms each
- 100% determinism verified (identical hashes)

+++ benches/regression.rs
@@ -17,7 +17,7 @@ fn bench_cpg_build(c: &mut Criterion) {
                     CPGNodeId(i),
                     CPGNodeKind::Function,
                     OriginRef::Function { function_id: semantic::model::FunctionId(i) },
-                    ByteRange::new(i as u32 * 10, (i as u32 + 1) * 10),
+                    ByteRange::new((i * 10) as usize, ((i + 1) * 10) as usize),
                 )));
             }
             cpg
@@ -32,7 +32,7 @@ fn bench_query_execution(c: &mut Criterion) {
             CPGNodeId(i),
             CPGNodeKind::Function,
             OriginRef::Function { function_id: semantic::model::FunctionId(i) },
-            ByteRange::new(i as u32 * 10, (i as u32 + 1) * 10),
+            ByteRange::new((i * 10) as usize, ((i + 1) * 10) as usize),
         ));
     }
 
@@ -61,7 +61,7 @@ fn bench_cpg_hash(c: &mut Criterion) {
             CPGNodeId(i),
             CPGNodeKind::Function,
             OriginRef::Function { function_id: semantic::model::FunctionId(i) },
-            ByteRange::new(i as u32 * 10, (i as u32 + 1) * 10),
+            ByteRange::new((i * 10) as usize, ((i + 1) * 10) as usize),
         ));
     }
